feat(seed): change seeder behavior & add dummy one for tesing purpose
