+++ packages/api/.env.test
@@ -1,0 +1,7 @@
+APP_ENV=test
+APP_SECRET=58DD3Xryhm7t7gREk2SUTBj2gup5UQIt
+APP_PREFORK=false
+
+TRACING_ENABLED=false
+
+DB_NAME=supateam_test
