+++ packages/api/go.mod
@@ -3,21 +3,25 @@ module github.com/marmorag/supateam
 go 1.15
 
 require (
-	github.com/JDR-ynovant/api v0.0.0-20210624205158-8b6961d23669
+	github.com/andybalholm/brotli v1.0.3 // indirect
 	github.com/arsmn/fiber-swagger/v2 v2.20.0
-	github.com/aschenmaker/fiber-opentracing v0.0.1 // indirect
-	github.com/dgrijalva/jwt-go v3.2.0+incompatible // indirect
-	github.com/getsentry/sentry-go v0.12.0 // indirect
-	github.com/go-playground/validator v9.31.0+incompatible // indirect
+	github.com/aschenmaker/fiber-opentracing v0.0.1
+	github.com/dgrijalva/jwt-go v3.2.0+incompatible
+	github.com/getsentry/sentry-go v0.12.0
+	github.com/go-openapi/swag v0.19.15 // indirect
+	github.com/go-playground/universal-translator v0.17.0 // indirect
+	github.com/go-playground/validator v9.31.0+incompatible
 	github.com/gofiber/fiber/v2 v2.21.0
-	github.com/gofiber/jwt/v2 v2.2.7 // indirect
-	github.com/golang-jwt/jwt/v4 v4.1.0 // indirect
-	github.com/google/uuid v1.3.0 // indirect
+	github.com/golang-jwt/jwt/v4 v4.1.0
+	github.com/google/uuid v1.3.0
 	github.com/joho/godotenv v1.4.0
-	github.com/opentracing/opentracing-go v1.2.0 // indirect
+	github.com/leodido/go-urn v1.2.1 // indirect
+	github.com/mailru/easyjson v0.7.7 // indirect
+	github.com/opentracing/opentracing-go v1.2.0
 	github.com/spf13/cobra v1.2.1
-	github.com/swaggo/swag v1.7.3 // indirect
+	github.com/swaggo/swag v1.7.3
 	go.mongodb.org/mongo-driver v1.7.3
-	golang.org/x/crypto v0.0.0-20210921155107-089bfa567519 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
+	golang.org/x/crypto v0.0.0-20210921155107-089bfa567519
+	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
+	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
 )
