+++ packages/api/internal/seeder/dummy_seeder.go
@@ -1,0 +1,11 @@
+package seeder
+
+import "fmt"
+
+type DummySeeder struct{}
+
+func (d DummySeeder) Seed() error {
+	fmt.Println("dummy seeder")
+
+	return nil
+}
